Do all post-processing in 1 step, fixes structures being cut off
