+++ src/main/java/com/gmail/sharpcastle33/did/generator/Centroid.java
@@ -1,0 +1,13 @@
+package com.gmail.sharpcastle33.did.generator;
+
+import com.sk89q.worldedit.math.Vector3;
+
+public class Centroid {
+    public final Vector3 pos;
+    public final int size;
+
+    public Centroid(Vector3 pos, int size) {
+        this.pos = pos;
+        this.size = size;
+    }
+}
